refactor: tx expression, query key,  and change file name
