+++ src/lib/query/index.ts
@@ -3,3 +3,4 @@ export * from "./code";
 export * from "./contract";
 export * from "./proposal";
 export * from "./tx";
+export * from "./account";
