+++ src/lib/services/accountService.ts
@@ -2,8 +2,8 @@ import { useWallet } from "@cosmos-kit/react";
 import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 
-import { useCelatoneApp, useChainId } from "lib/app-provider";
-import { getAccountIdByAddressQueryDocument } from "lib/query/account";
+import { useCelatoneApp } from "lib/app-provider";
+import { getAccountIdByAddressQueryDocument } from "lib/query";
 import type { Addr, Balance, Option } from "lib/types";
 
 import { getAccountBalanceInfo } from "./account";
@@ -34,17 +34,18 @@ export const useAccountBalances = (
   );
 };
 
-export const useAccountId = (walletAddress: Option<Addr>) => {
+export const useAccountId = (
+  walletAddress: Option<Addr>
+): UseQueryResult<Option<number>> => {
   const { indexerGraphClient } = useCelatoneApp();
-  const chainId = useChainId();
   const queryFn = () => {
     if (!walletAddress)
       throw new Error("Error fetching account id: failed to retrieve address.");
     return indexerGraphClient
       .request(getAccountIdByAddressQueryDocument, { address: walletAddress })
-      .then(({ accounts_by_pk }) => accounts_by_pk?.id);
+      .then<Option<number>>(({ accounts_by_pk }) => accounts_by_pk?.id);
   };
-  return useQuery(["account_id", chainId, walletAddress], queryFn, {
+  return useQuery(["account_id", indexerGraphClient, walletAddress], queryFn, {
     enabled: Boolean(walletAddress),
     retry: 1,
     refetchOnWindowFocus: false,
